fixes routing between home to account details fragment and transfer fragment, TODO - handle fragment state and back navigation
