+++ app/src/main/java/com/hover/stax/account/AccountsFragment.kt
@@ -13,9 +13,9 @@ import com.hover.stax.channels.Channel
 import com.hover.stax.channels.ChannelsRecyclerViewAdapter
 import com.hover.stax.channels.ChannelsViewModel
 import com.hover.stax.databinding.FragmentAccountsBinding
+import com.hover.stax.home.MainActivity
 import com.hover.stax.requests.RequestActivity
 import com.hover.stax.transfers.TransactionType
-import com.hover.stax.transfers.TransferActivity
 import com.hover.stax.utils.UIHelper
 import kotlinx.coroutines.launch
 import org.koin.androidx.viewmodel.ext.android.viewModel
@@ -71,7 +71,7 @@ class AccountsFragment : Fragment(), ChannelsRecyclerViewAdapter.SelectListener,
     }
 
     private fun fetchAccounts(action: HoverAction, channel: Channel) {
-        (activity as? TransferActivity)?.makeCall(action, channel)
+        (activity as? MainActivity)?.makeCall(action, channel)
                 ?: (activity as? RequestActivity)?.makeCall(action, channel)
     }
