+++ app/src/main/java/com/hover/stax/hover/TransactionReceiver.kt
@@ -118,10 +118,10 @@ class TransactionReceiver : BroadcastReceiver(), KoinComponent {
         while (matcher.find()) {
             val newAccount = Account(matcher.group(1), channel!!)
             accounts.add(newAccount)
-
-            if (repo.getDefaultAccount() == null && accounts.isNotEmpty())
-                accounts.first().isDefault = true
         }
+
+        if (repo.getDefaultAccount() == null && accounts.isNotEmpty())
+            accounts.first().isDefault = true
 
         Timber.e("Found ${accounts.size} accounts")
