+++ app/src/main/java/com/hover/stax/transfers/TransferActivity.kt
@@ -25,115 +25,115 @@ import timber.log.Timber
 
 class TransferActivity : AbstractNavigationActivity(), PushNotificationTopicsInterface {
 
-    private val actionSelectViewModel: ActionSelectViewModel by viewModel()
-    private val channelsViewModel: ChannelsViewModel by viewModel()
-    private val transferViewModel: TransferViewModel by viewModel()
-    private lateinit var scheduleViewModel: ScheduleDetailViewModel
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        intent.action?.let {
-            transferViewModel.setTransactionType(it)
-            channelsViewModel.setType(it)
-        }
-
-        checkIntent()
-        setContentView(R.layout.activity_transfer)
-
-        setUpNav()
-    }
-
-    private fun checkIntent() {
-        when {
-            intent.hasExtra(Schedule.SCHEDULE_ID) -> createFromSchedule(intent.getIntExtra(Schedule.SCHEDULE_ID, -1))
-            intent.hasExtra(Constants.REQUEST_LINK) -> createFromRequest(intent.getStringExtra(Constants.REQUEST_LINK)!!)
-            else -> Utils.logAnalyticsEvent(getString(R.string.visit_screen, intent.action), this)
-        }
-    }
-
-    private fun createFromSchedule(scheduleId: Int) {
-        scheduleViewModel = getViewModel()
-        with(scheduleViewModel) {
-            action.observe(this@TransferActivity, { it?.let { actionSelectViewModel.setActiveAction(it) } })
-            schedule.observe(this@TransferActivity, { it?.let { transferViewModel.view(it) } })
-            setSchedule(scheduleId)
-        }
-
-        Utils.logAnalyticsEvent(getString(R.string.clicked_schedule_notification), this)
-    }
-
-    private fun createFromRequest(link: String) {
-        transferViewModel.decrypt(link)
-        observeRequest()
-        Utils.logAnalyticsEvent(getString(R.string.clicked_request_link), this)
-    }
-
-    private fun observeRequest() {
-        val alertDialog = StaxDialog(this).setDialogMessage(R.string.loading_link_dialoghead).showIt()
-        transferViewModel.request.observe(this@TransferActivity, { it?.let { alertDialog?.dismiss() } })
-    }
-
-    fun submit(account: Account) = actionSelectViewModel.activeAction.value?.let { makeHoverCall(it, account) }
-
-    private fun makeHoverCall(action: HoverAction, account: Account) {
-        Utils.logAnalyticsEvent(getString(R.string.finish_transfer, TransactionType.type), this)
-        updatePushNotifGroupStatus()
-
-        transferViewModel.checkSchedule()
-
-        makeCall(action, selectedAccount = account)
-    }
-
-    private fun getRequestCode(transactionType: String) : Int {
-        return if(transactionType == HoverAction.FETCH_ACCOUNTS) Constants.FETCH_ACCOUNT_REQUEST
-        else Constants.TRANSFER_REQUEST
-    }
-
-    fun makeCall(action: HoverAction, channel: Channel? = null, selectedAccount: Account? = null) {
-        val hsb = HoverSession.Builder(action, channel
-                ?: channelsViewModel.activeChannel.value!!, this, getRequestCode(action.transaction_type))
-
-        if (action.transaction_type != HoverAction.FETCH_ACCOUNTS) {
-            hsb.extra(HoverAction.AMOUNT_KEY, transferViewModel.amount.value)
-                    .extra(HoverAction.NOTE_KEY, transferViewModel.note.value)
-                    .extra(Constants.ACCOUNT_NAME, selectedAccount?.name)
-
-            selectedAccount?.run { hsb.setAccountId(id.toString()) }
-            transferViewModel.contact.value?.let { addRecipientInfo(hsb) }
-        }
-
-        hsb.run()
-    }
-
-    private fun addRecipientInfo(hsb: HoverSession.Builder) {
-        hsb.extra(HoverAction.ACCOUNT_KEY, transferViewModel.contact.value!!.accountNumber)
-                .extra(
-                        HoverAction.PHONE_KEY, PhoneHelper.getNumberFormatForInput(
-                        transferViewModel.contact.value?.accountNumber,
-                        actionSelectViewModel.activeAction.value, channelsViewModel.activeChannel.value
-                )
-                )
-    }
-
-    private fun updatePushNotifGroupStatus() {
-        joinTransactionGroup(this)
-        leaveNoUsageGroup(this)
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == Constants.TRANSFER_REQUEST)
-            returnResult(requestCode, resultCode, data)
-    }
-
-    private fun returnResult(type: Int, result: Int, data: Intent?) {
-        val i = data?.let { Intent(it) } ?: Intent()
-        transferViewModel.contact.value?.let { i.putExtra(StaxContact.ID_KEY, it.lookupKey) }
-        i.action = if (type == Constants.SCHEDULE_REQUEST) Constants.SCHEDULED else Constants.TRANSFERRED
-        setResult(result, i)
-        finish()
-    }
-
-    override fun onBackPressed() = if (transferViewModel.isEditing.value == false) transferViewModel.setEditing(true) else super.onBackPressed()
+//    private val actionSelectViewModel: ActionSelectViewModel by viewModel()
+//    private val channelsViewModel: ChannelsViewModel by viewModel()
+//    private val transferViewModel: TransferViewModel by viewModel()
+//    private lateinit var scheduleViewModel: ScheduleDetailViewModel
+//
+//    override fun onCreate(savedInstanceState: Bundle?) {
+//        super.onCreate(savedInstanceState)
+//
+//        intent.action?.let {
+//            transferViewModel.setTransactionType(it)
+//            channelsViewModel.setType(it)
+//        }
+//
+//        checkIntent()
+//        setContentView(R.layout.activity_transfer)
+//
+//        setUpNav()
+//    }
+//
+//    private fun checkIntent() {
+//        when {
+//            intent.hasExtra(Schedule.SCHEDULE_ID) -> createFromSchedule(intent.getIntExtra(Schedule.SCHEDULE_ID, -1))
+//            intent.hasExtra(Constants.REQUEST_LINK) -> createFromRequest(intent.getStringExtra(Constants.REQUEST_LINK)!!)
+//            else -> Utils.logAnalyticsEvent(getString(R.string.visit_screen, intent.action), this)
+//        }
+//    }
+//
+//    private fun createFromSchedule(scheduleId: Int) {
+//        scheduleViewModel = getViewModel()
+//        with(scheduleViewModel) {
+//            action.observe(this@TransferActivity, { it?.let { actionSelectViewModel.setActiveAction(it) } })
+//            schedule.observe(this@TransferActivity, { it?.let { transferViewModel.view(it) } })
+//            setSchedule(scheduleId)
+//        }
+//
+//        Utils.logAnalyticsEvent(getString(R.string.clicked_schedule_notification), this)
+//    }
+//
+//    private fun createFromRequest(link: String) {
+//        transferViewModel.decrypt(link)
+//        observeRequest()
+//        Utils.logAnalyticsEvent(getString(R.string.clicked_request_link), this)
+//    }
+//
+//    private fun observeRequest() {
+//        val alertDialog = StaxDialog(this).setDialogMessage(R.string.loading_link_dialoghead).showIt()
+//        transferViewModel.request.observe(this@TransferActivity, { it?.let { alertDialog?.dismiss() } })
+//    }
+//
+//    fun submit(account: Account) = actionSelectViewModel.activeAction.value?.let { makeHoverCall(it, account) }
+//
+//    private fun makeHoverCall(action: HoverAction, account: Account) {
+//        Utils.logAnalyticsEvent(getString(R.string.finish_transfer, TransactionType.type), this)
+//        updatePushNotifGroupStatus()
+//
+//        transferViewModel.checkSchedule()
+//
+//        makeCall(action, selectedAccount = account)
+//    }
+//
+//    private fun getRequestCode(transactionType: String) : Int {
+//        return if(transactionType == HoverAction.FETCH_ACCOUNTS) Constants.FETCH_ACCOUNT_REQUEST
+//        else Constants.TRANSFER_REQUEST
+//    }
+//
+//    fun makeCall(action: HoverAction, channel: Channel? = null, selectedAccount: Account? = null) {
+//        val hsb = HoverSession.Builder(action, channel
+//                ?: channelsViewModel.activeChannel.value!!, this, getRequestCode(action.transaction_type))
+//
+//        if (action.transaction_type != HoverAction.FETCH_ACCOUNTS) {
+//            hsb.extra(HoverAction.AMOUNT_KEY, transferViewModel.amount.value)
+//                    .extra(HoverAction.NOTE_KEY, transferViewModel.note.value)
+//                    .extra(Constants.ACCOUNT_NAME, selectedAccount?.name)
+//
+//            selectedAccount?.run { hsb.setAccountId(id.toString()) }
+//            transferViewModel.contact.value?.let { addRecipientInfo(hsb) }
+//        }
+//
+//        hsb.run()
+//    }
+//
+//    private fun addRecipientInfo(hsb: HoverSession.Builder) {
+//        hsb.extra(HoverAction.ACCOUNT_KEY, transferViewModel.contact.value!!.accountNumber)
+//                .extra(
+//                        HoverAction.PHONE_KEY, PhoneHelper.getNumberFormatForInput(
+//                        transferViewModel.contact.value?.accountNumber,
+//                        actionSelectViewModel.activeAction.value, channelsViewModel.activeChannel.value
+//                )
+//                )
+//    }
+//
+//    private fun updatePushNotifGroupStatus() {
+//        joinTransactionGroup(this)
+//        leaveNoUsageGroup(this)
+//    }
+//
+//    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+//        super.onActivityResult(requestCode, resultCode, data)
+//        if (requestCode == Constants.TRANSFER_REQUEST)
+//            returnResult(requestCode, resultCode, data)
+//    }
+//
+//    private fun returnResult(type: Int, result: Int, data: Intent?) {
+//        val i = data?.let { Intent(it) } ?: Intent()
+//        transferViewModel.contact.value?.let { i.putExtra(StaxContact.ID_KEY, it.lookupKey) }
+//        i.action = if (type == Constants.SCHEDULE_REQUEST) Constants.SCHEDULED else Constants.TRANSFERRED
+//        setResult(result, i)
+//        finish()
+//    }
+//
+//    override fun onBackPressed() = if (transferViewModel.isEditing.value == false) transferViewModel.setEditing(true) else super.onBackPressed()
 }
