+++ app/src/main/res/layout/fragment_channel.xml
@@ -44,7 +44,7 @@
                 <TableRow android:layout_marginVertical="@dimen/margin_21">
 
                     <TextView
-                            android:layout_weight="1"
+                            android:layout_weight="2"
                             android:fontFamily="font-normal"
                             android:text="@string/moneyout_label"
                             android:textColor="@color/offWhite"
@@ -56,6 +56,7 @@
                             android:textAlignment="viewEnd"
                             android:textColor="@color/offWhite"
                             android:textSize="@dimen/text_21"
+                            android:layout_weight="1"
                             android:textStyle="bold" />
                 </TableRow>
 
@@ -64,13 +65,14 @@
                         android:visibility="gone">
 
                     <TextView
-                            android:layout_weight="1"
+                            android:layout_weight="2"
                             android:fontFamily="font-normal"
                             android:text="@string/moneyin_label"
                             android:textColor="@color/offWhite"
                             android:textSize="@dimen/text_16" />
 
                     <TextView
+                            android:layout_weight="1"
                             android:id="@+id/details_money_in"
                             android:fontFamily="font-normal"
                             android:textAlignment="viewEnd"
@@ -83,7 +85,7 @@
 
                     <TextView
                             android:id="@+id/fees_description"
-                            android:layout_weight="1"
+                            android:layout_weight="2"
                             android:fontFamily="font-normal"
                             android:text="@string/fees_label"
                             android:textColor="@color/offWhite"
@@ -92,6 +94,7 @@
                     <TextView
                             android:id="@+id/details_fees"
                             android:fontFamily="font-normal"
+                            android:layout_weight="1"
                             android:textAlignment="viewEnd"
                             android:textColor="@color/offWhite"
                             android:textSize="@dimen/text_21"
@@ -102,7 +105,7 @@
 
                     <TextView
                             android:id="@+id/orig_name_desc"
-                            android:layout_weight="1"
+                            android:layout_weight="2"
                             android:fontFamily="font-normal"
                             android:text="@string/original_name"
                             android:textColor="@color/offWhite"
@@ -111,6 +114,7 @@
                     <TextView
                             android:id="@+id/original_name"
                             android:fontFamily="font-normal"
+                            android:layout_weight="1"
                             android:textAlignment="viewEnd"
                             android:textColor="@color/offWhite"
                             android:textSize="@dimen/text_21"
